Changed tracker page according to our design
Removed rest tracking feature and reps feature; we want the user to focus on exercising first. Necessary features left are the timer (in text and in circular form) and exercise name. Added gamification features: plant in the middle of the timer and plant cheering for the user as they exercise.